requirements.txt
streamlit
google.oauth2
googleapiclient.discovery
googleapiclient.errors
google.oauth2.credentials